Extracted logic from RequestBuilderFactory to collaborators
